Finished test for #239.
Accidentally did _too_ good of job replacing the '?' SVG and changed one of them back to the question mark (in *word-entries.html*).

Wrote the proper test for #239 in *search.spec.js* and it passed 🎉. If all is good, I think this issue is ready for review and/or to be merged.