Refactor search, part 2
Building on the previous move into separate files, this introduces more
classes and objects. The general workflow is:

  - Search methods generate candidate Result objects, recording the
    features that indicate why a given Result may be a good match for the
    query

  - Features for the same results are aggregated together and can be used
    for ranking. If *multiple* search methods select a given result, that
    can be used as a signal to prioritize that result.

  - Presentation features like relabelling are handled separately in a
    later phase of the process

Again, user-facing changes should be minimal at this time. This is just
rearranging stuff so that there is room to add new functionality later.

The main outstanding issue that I am aware of at this point is that
changing the search result serialization may break click-in-text, in which
case I will have to add legacy compatibility keys.