Removed nb_black and moved to dawg2 dependencies
The dawg repo has not been updated in years, a fork continues
development for newer python versions.  see
https://github.com/pymorphy2-fork/DAWG/.

The nb_black repo has also not been updated, and it breaks compilation.
It is only needed for jupyter notebooks, to force Black code formatting.
An alternative approach should be followed, in the end, jupyter
notebooks should not be supported directly by this repository.